+++ polaris-react/src/components/BlockStack/tests/BlockStack.test.tsx
@@ -25,7 +25,7 @@ describe('<BlockStack />', () => {
 
   it('overrides custom properties if they are passed in', () => {
     const blockStack = mountWithApp(
-      <BlockStack inlineAlign="center" gap="10">
+      <BlockStack inlineAlign="center" gap="1000">
         {children}
       </BlockStack>,
     );
@@ -33,20 +33,20 @@ describe('<BlockStack />', () => {
     expect(blockStack).toContainReactComponent('div', {
       style: expect.objectContaining({
         '--pc-block-stack-inline-align': 'center',
-        '--pc-block-stack-gap-xs': 'var(--p-space-10)',
+        '--pc-block-stack-gap-xs': 'var(--p-space-1000)',
       }) as React.CSSProperties,
     });
   });
 
   it('accepts gap based on breakpoints', () => {
     const blockStack = mountWithApp(
-      <BlockStack gap={{xs: '2', md: '8'}}>{children}</BlockStack>,
+      <BlockStack gap={{xs: '200', md: '800'}}>{children}</BlockStack>,
     );
 
     expect(blockStack).toContainReactComponent('div', {
       style: expect.objectContaining({
-        '--pc-block-stack-gap-md': 'var(--p-space-8)',
-        '--pc-block-stack-gap-xs': 'var(--p-space-2)',
+        '--pc-block-stack-gap-md': 'var(--p-space-800)',
+        '--pc-block-stack-gap-xs': 'var(--p-space-200)',
       }) as React.CSSProperties,
     });
   });