+++ polaris-react/src/components/InlineGrid/InlineGrid.tsx
@@ -20,14 +20,14 @@ export interface InlineGridProps extends React.AriaAttributes {
   children?: React.ReactNode;
   /** The number of columns to display. Accepts either a single value or an object of values for different screen sizes.
    * @example
-   * columns={6}
-   * columns={{xs: 1, sm: 1, md: 3, lg: 6, xl: 6}}
+   * columns={600}
+   * columns={{xs: 100, sm: 100, md: 300, lg: 600, xl: 600}}
    */
   columns?: Columns;
   /** The spacing between children. Accepts a spacing token or an object of spacing tokens for different screen sizes.
    * @example
-   * gap='2'
-   * gap={{xs: '1', sm: '2', md: '3', lg: '4', xl: '5'}}
+   * gap='200'
+   * gap={{xs: '100', sm: '200', md: '300', lg: '400', xl: '500'}}
    */
   gap?: Gap;
   /** Vertical alignment of children. If not set, inline elements will stretch to the height of the parent.