+++ polaris-react/src/utilities/tests/css.test.ts
@@ -27,47 +27,47 @@ describe('variationName', () => {
 
 describe('getResponsiveProps', () => {
   it('takes a string and returns the custom property', () => {
-    expect(getResponsiveProps('stack', 'space', 'space', '4')).toMatchObject({
-      '--pc-stack-space-xs': 'var(--p-space-4)',
+    expect(getResponsiveProps('stack', 'space', 'space', '400')).toMatchObject({
+      '--pc-stack-space-xs': 'var(--p-space-400)',
     });
   });
 
   it('takes an object with a breakpoint and value and returns the property for each breakpoint', () => {
     expect(
-      getResponsiveProps('stack', 'space', 'space', {xs: '2', md: '8'}),
+      getResponsiveProps('stack', 'space', 'space', {xs: '200', md: '800'}),
     ).toMatchObject({
-      '--pc-stack-space-xs': 'var(--p-space-2)',
-      '--pc-stack-space-md': 'var(--p-space-8)',
+      '--pc-stack-space-xs': 'var(--p-space-200)',
+      '--pc-stack-space-md': 'var(--p-space-800)',
     });
   });
 
   it('handles a full responsive object', () => {
     expect(
       getResponsiveProps('stack', 'space', 'space', {
-        xs: '2',
-        sm: '4',
-        md: '8',
-        lg: '8',
-        xl: '10',
+        xs: '200',
+        sm: '400',
+        md: '800',
+        lg: '800',
+        xl: '1000',
       }),
     ).toMatchObject({
-      '--pc-stack-space-xs': 'var(--p-space-2)',
-      '--pc-stack-space-sm': 'var(--p-space-4)',
-      '--pc-stack-space-md': 'var(--p-space-8)',
-      '--pc-stack-space-lg': 'var(--p-space-8)',
-      '--pc-stack-space-xl': 'var(--p-space-10)',
+      '--pc-stack-space-xs': 'var(--p-space-200)',
+      '--pc-stack-space-sm': 'var(--p-space-400)',
+      '--pc-stack-space-md': 'var(--p-space-800)',
+      '--pc-stack-space-lg': 'var(--p-space-800)',
+      '--pc-stack-space-xl': 'var(--p-space-1000)',
     });
   });
 
   it('does not fill in leading undefined values', () => {
     expect(
       getResponsiveProps('stack', 'space', 'space', {
-        md: '4',
-        xl: '10',
+        md: '400',
+        xl: '1000',
       }),
     ).toMatchObject({
-      '--pc-stack-space-md': 'var(--p-space-4)',
-      '--pc-stack-space-xl': 'var(--p-space-10)',
+      '--pc-stack-space-md': 'var(--p-space-400)',
+      '--pc-stack-space-xl': 'var(--p-space-1000)',
     });
   });
 
@@ -75,11 +75,11 @@ describe('getResponsiveProps', () => {
     expect(
       getResponsiveProps<string | number>('stack', 'space', 'space', {
         md: 0,
-        xl: '10',
+        xl: '1000',
       }),
     ).toMatchObject({
       '--pc-stack-space-md': 'var(--p-space-0)',
-      '--pc-stack-space-xl': 'var(--p-space-10)',
+      '--pc-stack-space-xl': 'var(--p-space-1000)',
     });
   });
 
@@ -87,10 +87,10 @@ describe('getResponsiveProps', () => {
     expect(
       getResponsiveProps('stack', 'space', 'space', {
         md: undefined,
-        lg: '10',
+        lg: '1000',
       }),
     ).toMatchObject({
-      '--pc-stack-space-lg': 'var(--p-space-10)',
+      '--pc-stack-space-lg': 'var(--p-space-1000)',
     });
   });
 });